getEntrance

+++ smart-contract-lottery/contracts/Lottery.sol
@@ -39,9 +39,12 @@ contract Lottery is Ownable, VRFConsumerBase{
     }
 
     function getEntranceFee() public view returns (uint256){
-        (,int price,,,) = ethUsdPriceFeed.latestRoundData();
-        uint256 adjustedPrice = uint256(price) * 10**10;  // 18 decimals
-        uint256 costToEnter = (usdEntryFee * 10 ** 18) / adjustedPrice;
+        (, int256 price, , , ) = ethUsdPriceFeed.latestRoundData();
+        uint256 adjustedPrice = uint256(price) * 10**10; // 18 decimals
+        // $50, $2,000 / ETH
+        // 50/2,000
+        // 50 * 100000 / 2000
+        uint256 costToEnter = (usdEntryFee * 10**18) / adjustedPrice;
         return costToEnter;
        
     }
